Limit number of audio elements created.
Because the check for existing elements was faulty, endless
audio tags are created over time. In SPA or such this could be a lot
of elements. Got Chrome to AW SNAP at me after 15 minutes of
calling playAudio on an interval.

This fixes the check for existing elements.

You can call playaudio as often as you like, only ever creating just enough elements to play N simultaneous sounds.